test(testing): split eval/generate-dataset entrypoint tests into cfg-only modules
Mirror tests/test_train.py: keep the canonical tests/test_eval.py and
tests/test_generate_dataset.py to cfg-entrypoint tests only, and relocate the
helper/unit tests to clearly-named sibling modules.

eval: move the postprocessing argv tests, metric-IO tests, and the
_maybe_upload_output_dir + CLI-e2e tests out of tests/test_eval.py into
test_eval_postprocessing.py, test_eval_metrics.py, and test_eval_upload.py.

generate-dataset: promote test_generate_dataset_shards.py to the canonical
tests/test_generate_dataset.py (from_hydra(cfg_dataset) entrypoint tests), and
rename the direct-call generate()/main() suite to
tests/pipeline/entrypoints/test_generate_dataset_unit.py (content unchanged).

Add tests/_meta/test_entrypoint_test_modules.py: fails if a canonical
entrypoint test module imports a private synth_setter.cli helper, so the
cfg-only split can't silently erode; each module docstring carries a matching
nudge.

Update the generate-dataset-shards GHA workflow, docs/reference/testing.md, the
helper comment refs, and repath the renamed file's existing
.pydoclint-baseline.txt entry. The collected test-node set is unchanged.

Refs #1400
